pyproject.toml
[tool.poetry]
name = "scoredraft"
version = "0.1.0"
description = "Light-weight music sketchbook"
authors = ["Stéfano Troncaro <36011367+stroncaro@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.3"
mypy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
